+++ ServerDevcommands/Commands/PlayerList.cs
@@ -1,0 +1,32 @@
+using System;
+using System.Linq;
+
+namespace ServerDevcommands;
+///<summary>Prints players and their ids.</summary>
+public class PlayerListCommand
+{
+  private string Format(ZNetPeer player)
+  {
+    return String.Format(Settings.Format(Settings.PlayerListFormat),
+      player.m_playerName, player.m_socket.GetHostName(), player.m_characterID.m_userID, player.m_refPos.x, player.m_refPos.y, player.m_refPos.z
+    );
+  }
+  private string Format()
+  {
+    var pos = ZNet.instance.m_referencePosition;
+    return String.Format(Settings.Format(Settings.PlayerListFormat),
+      Game.instance.GetPlayerProfile().GetName(), PrivilegeManager.GetNetworkUserId(), ZNet.instance.m_characterID, pos.x, pos.y, pos.z
+    );
+  }
+  public PlayerListCommand()
+  {
+    Helper.Command("playerlist", "- Prints online players.", (args) =>
+    {
+      var players = ZNet.instance.GetPeers().Select(Format).ToList();
+      if (ZNet.instance && !ZNet.instance.IsDedicated())
+        players.Add(Format());
+      args.Context.AddString(string.Join("\n", players));
+    });
+    AutoComplete.RegisterEmpty("playerlist");
+  }
+}